+++ src/main/java/palamod/procedures/Truetresure_dropProcedure.java
@@ -10,9 +10,9 @@ import net.minecraft.world.entity.item.ItemEntity;
 
 public class Truetresure_dropProcedure {
 	public static void execute(LevelAccessor world, double x, double y, double z) {
-		for (int index0 = 0; index0 < (int) (3); index0++) {
+		for (int index0 = 0; index0 < 3; index0++) {
 			if (Math.random() < 0.7) {
-				for (int index1 = 0; index1 < (int) (8); index1++) {
+				for (int index1 = 0; index1 < 8; index1++) {
 					if (world instanceof Level _level && !_level.isClientSide()) {
 						ItemEntity entityToSpawn = new ItemEntity(_level, x, y, z, new ItemStack(PalamodModItems.PALADIUM_INGOT.get()));
 						entityToSpawn.setPickUpDelay(10);
@@ -33,7 +33,7 @@ public class Truetresure_dropProcedure {
 				}
 			}
 			if (Math.random() < 0.25) {
-				for (int index2 = 0; index2 < (int) (4); index2++) {
+				for (int index2 = 0; index2 < 4; index2++) {
 					if (world instanceof Level _level && !_level.isClientSide()) {
 						ItemEntity entityToSpawn = new ItemEntity(_level, x, y, z, new ItemStack(PalamodModItems.ORANGEBLUE.get()));
 						entityToSpawn.setPickUpDelay(10);
